Changing the RAP call according to the spec
The frame handling is in fact quite tricky. Sharing and mutation are
very important in a (DUM LDF* RAP) sequence, which is the way of
building mutual recursion -- instead of having an explicit "recursive
closure construction" instruction.

The specification has changed during the night, and this new
implementation should be closer to the new spec.